Factories y Seeders agregados

+++ database/migrations/2021_08_19_183626_create_asesores_table.php
@@ -15,7 +15,7 @@ class CreateAsesoresTable extends Migration
     {
         Schema::create('asesores', function (Blueprint $table) {
             $table->id();
-            $table->smallInteger('cip');
+            $table->integer('cip');
             $table->unsignedBigInteger('docente_id')->nullable();
             $table->foreign('docente_id')->references('id')->on('docentes');
         });
