+++ database/migrations/2021_08_19_183724_create_jurados_table.php
@@ -15,7 +15,7 @@ class CreateJuradosTable extends Migration
     {
         Schema::create('jurados', function (Blueprint $table) {
             $table->id();
-            $table->smallInteger('cip')->nullable();
+            $table->integer('cip')->nullable();
             $table->unsignedBigInteger('docente_id')->nullable();
             $table->foreign('docente_id')->references('id')->on('docentes');
         });
